autostart change

+++ home/.local/share/dwm/autostart.sh
@@ -2,4 +2,17 @@
 asnddef &
 mon &
 sleep 5 && ~/.fehbg &
-pkill sunshine ; sunshine &
+
+
+if command -v sunshine >/dev/null 2>&1; then
+    pkill sunshine ; sunshine &
+fi
+
+if command -v gshorts >/dev/null 2>&1; then
+    pkill gshorts ; gshorts &
+fi
+
+if command -v slack >/dev/null 2>&1; then
+    slack &
+    echo slack started
+fi
